added endpoint for actually running the script
